pyproject.toml
[build-system]
requires = ["setuptools>=61", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "task2-baseline"
version = "0.1.0"
description = "PyTorch baseline for ICASSP 2026 Task 2"
requires-python = ">=3.9"
dependencies = [
  "numpy",
  "pandas",
  "torch",
  "torchvision",
  "Pillow",
  "tqdm",
]

[tool.setuptools]
package-dir = {"" = "."}

[tool.setuptools.packages.find]
where = ["."]
include = ["task2_baseline*"]
